sheets/requirements.txt
google-api-python-client==2.167.0
google-auth-httplib2==0.2.0
gspread==6.2.0
pytest==8.3.5
